+++ R/haven.R
@@ -13,9 +13,10 @@ NULL
 #'   processed with \code{\link[readr]{datasource}()}.
 #' @param data Data frame to write.
 #' @param path Path to file where the data will be written.
-#' @param encoding The character encoding used for the file. This defaults to
-#'   the encoding specified in the file, or UTF-8. You can use this argument
-#'   to override the value stored in the file if it is correct
+#' @param encoding,catalog_encoding The character encoding used for the
+#'   `data_file` and `catalog_encoding` respectively. A value of `NULL`
+#'   uses the encoding specified in the file; use this argument to override it
+#'   if it is incorrect.
 #' @param cols_only A character vector giving an experimental way to read in
 #'   only specified columns.
 #' @return A tibble, data frame variant with nice defaults.
@@ -26,7 +27,8 @@ NULL
 #' @examples
 #' path <- system.file("examples", "iris.sas7bdat", package = "haven")
 #' read_sas(path)
-read_sas <- function(data_file, catalog_file = NULL, encoding = NULL,
+read_sas <- function(data_file, catalog_file = NULL,
+                     encoding = NULL, catalog_encoding = encoding,
                      cols_only = NULL) {
   if (is.null(encoding)) {
     encoding <- ""
@@ -43,8 +45,8 @@ read_sas <- function(data_file, catalog_file = NULL, encoding = NULL,
   }
 
   switch(class(spec_data)[1],
-    source_file = df_parse_sas_file(spec_data, spec_cat, encoding = encoding, cols_only = cols_only),
-    source_raw = df_parse_sas_raw(spec_data, spec_cat, encoding = encoding, cols_only = cols_only),
+    source_file = df_parse_sas_file(spec_data, spec_cat, encoding = encoding, catalog_encoding = catalog_encoding, cols_only = cols_only),
+    source_raw = df_parse_sas_raw(spec_data, spec_cat, encoding = encoding, catalog_encoding = catalog_encoding, cols_only = cols_only),
     stop("This kind of input is not handled", call. = FALSE)
   )
 }